+++ src/components/checkmark-reset/index.ts
@@ -1,0 +1,1 @@
+export { CheckmarkResetButton } from './CheckmarkResetButton'